requirements.txt
imageio
pygame
python-telegram-bot
